Add javadoc comment for classes
