--- data/carData.txt
@@ -1,10 +1,0 @@
-Toyota Corolla | SGM4932K | 120.0 | false
-Toyota Corolla | SGM4932K | 120.0 | true
-Toyota Corolla | SGM4932K | 120.0 | false
-Toyota Corolla | SGM4932K | 120.0 | true
-Toyota Corolla | SGM4932K | 120.0 | false
-Toyota Corolla | SGM4932K | 120.0 | true
-Toyota Corolla | SGM4932K | 100.0 | false
-Toyota Corolla | SGM4932K | 90.0 | false
-Toyota Corolla | SGM4932K | 10.0 | false
-Toyolla | 32K | 100.0 | false
