--- data/customerData.txt
@@ -1,6 +1,0 @@
-John | 22 | 90907638
-John | 22 | 90907638
-John | 22 | 90907638
-John | 22 | 90907638
-John | 22 | 90907638
-John | 22 | 90907638
