+++ text-ui-test/testingCommands
@@ -5,7 +5,7 @@ remove-user /u John
 add-car /n Toyota Corolla /c SGM4932K /p 120
 list-cars
 remove-car /i SGM4932K
-add-tx /l SGT3238B /b John /d 14 /c 2024-10-10
+add-tx /p SBS123B /u John /d 30 /s 2024-10-17
 list-tx
 remove-tx /t 1
 exit
